Made some enemy and progression changes
Enemies are now slower to make it easier on players. Enemies are now 1 hit kills, so it's quicker to kill them, meaning the number of enemies don't stack up as much. Enemies are knocked over with physics when hit, and are de-spawned after 3 seconds. The player can pick them up in this time throw them to damage other enemies. The de-spawn timer is stopped when picking them up, and reset when they're put down or thrown. The buttons for the puzzles now work, and can only be pressed once all enemies in the level have been killed. The button is required to activate the trigger allowing players to move on to the next level.